@@ -1,83 +1,74 @@
 class LibrariesController < ApplicationController
+  before_action :set_library, only: [:show, :edit, :update, :destroy]
+
   # GET /libraries
-  # GET /libraries.xml
+  # GET /libraries.json
   def index
     @libraries = Library.all
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.xml  { render :xml => @libraries }
-    end
   end
 
   # GET /libraries/1
-  # GET /libraries/1.xml
+  # GET /libraries/1.json
   def show
-    @library = Library.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.xml  { render :xml => @library }
-    end
   end
 
   # GET /libraries/new
-  # GET /libraries/new.xml
   def new
     @library = Library.new
-
-    respond_to do |format|
-      format.html # new.html.erb
-      format.xml  { render :xml => @library }
-    end
   end
 
   # GET /libraries/1/edit
   def edit
-    @library = Library.find(params[:id])
   end
 
   # POST /libraries
-  # POST /libraries.xml
+  # POST /libraries.json
   def create
-    @library = Library.new(params[:library])
+    @library = Library.new(library_params)
 
     respond_to do |format|
       if @library.save
-        format.html { redirect_to(@library, :notice => 'Library was successfully created.') }
-        format.xml  { render :xml => @library, :status => :created, :location => @library }
+        format.html { redirect_to @library, notice: 'Library was successfully created.' }
+        format.json { render action: 'show', status: :created, location: @library }
       else
-        format.html { render :action => "new" }
-        format.xml  { render :xml => @library.errors, :status => :unprocessable_entity }
+        format.html { render action: 'new' }
+        format.json { render json: @library.errors, status: :unprocessable_entity }
       end
     end
   end
 
-  # PUT /libraries/1
-  # PUT /libraries/1.xml
+  # PATCH/PUT /libraries/1
+  # PATCH/PUT /libraries/1.json
   def update
-    @library = Library.find(params[:id])
-
     respond_to do |format|
-      if @library.update_attributes(params[:library])
-        format.html { redirect_to(@library, :notice => 'Library was successfully updated.') }
-        format.xml  { head :ok }
+      if @library.update(library_params)
+        format.html { redirect_to @library, notice: 'Library was successfully updated.' }
+        format.json { head :no_content }
       else
-        format.html { render :action => "edit" }
-        format.xml  { render :xml => @library.errors, :status => :unprocessable_entity }
+        format.html { render action: 'edit' }
+        format.json { render json: @library.errors, status: :unprocessable_entity }
       end
     end
   end
 
   # DELETE /libraries/1
-  # DELETE /libraries/1.xml
+  # DELETE /libraries/1.json
   def destroy
-    @library = Library.find(params[:id])
     @library.destroy
-
     respond_to do |format|
-      format.html { redirect_to(libraries_url) }
-      format.xml  { head :ok }
+      format.html { redirect_to libraries_url }
+      format.json { head :no_content }
     end
   end
+
+  private
+    # Use callbacks to share common setup or constraints between actions.
+    def set_library
+      @library = Library.find(params[:id])
+    end
+
+    # Never trust parameters from the scary internet, only allow the white list through.
+    def library_params
+      params[:library]
+    end
 end
