@@ -1,101 +1,90 @@
 class MixesController < ApplicationController
+  before_action :set_mix, only: [:add_sample, :show, :edit, :update, :destroy, :tracker]
+  before_action :authenticate_musician!, except: [:list, :index, :show]
+
   def add_sample
     sample_id = params[:sample_id].to_i
     
     mix_id = params[:id]
-    importage = Importage.create!(:sample_id => sample_id,
-                                 
-                                 :mix_id => mix_id)
-  respond_to do |format|
-    format.json { render :json => importage}
-  end
+    importage = Importage.create!(sample_id: sample_id, mix_id: mix_id)
+    respond_to do |format|
+      format.json { render json: importage }
+    end
   end
   
   def tracker
-    @mix = Mix.find params[:id]
     @tracks = @mix.samples.uniq
     @xToTimeConst = 10
   end
-  
+
   # GET /mixes
-  # GET /mixes.xml
+  # GET /mixes.json
   def index
     @mixes = Mix.all
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.xml  { render :xml => @mixes }
-    end
   end
 
   # GET /mixes/1
-  # GET /mixes/1.xml
+  # GET /mixes/1.json
   def show
-    @mix = Mix.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.xml  { render :xml => @mix }
-    end
   end
 
   # GET /mixes/new
-  # GET /mixes/new.xml
   def new
     @mix = Mix.new
-    @mix.musician = current_musician
-    respond_to do |format|
-      format.html # new.html.erb
-      format.xml  { render :xml => @mix }
-    end
   end
 
   # GET /mixes/1/edit
   def edit
-    @mix = Mix.find(params[:id])
   end
 
   # POST /mixes
-  # POST /mixes.xml
+  # POST /mixes.json
   def create
-    @mix = Mix.new(params[:mix])
+    @mix = Mix.new(mix_params().merge({ musician_id: current_musician.id}))
 
     respond_to do |format|
       if @mix.save
-        format.html { redirect_to(@mix, :notice => 'Mix was successfully created.') }
-        format.xml  { render :xml => @mix, :status => :created, :location => @mix }
+        format.html { redirect_to @mix, notice: 'Mix was successfully created.' }
+        format.json { render action: 'show', status: :created, location: @mix }
       else
-        format.html { render :action => "new" }
-        format.xml  { render :xml => @mix.errors, :status => :unprocessable_entity }
+        format.html { render action: 'new' }
+        format.json { render json: @mix.errors, status: :unprocessable_entity }
       end
     end
   end
 
-  # PUT /mixes/1
-  # PUT /mixes/1.xml
+  # PATCH/PUT /mixes/1
+  # PATCH/PUT /mixes/1.json
   def update
-    @mix = Mix.find(params[:id])
-
     respond_to do |format|
-      if @mix.update_attributes(params[:mix])
-        format.html { redirect_to(@mix, :notice => 'Mix was successfully updated.') }
-        format.xml  { head :ok }
+      if @mix.update(mix_params)
+        format.html { redirect_to @mix, notice: 'Mix was successfully updated.' }
+        format.json { head :no_content }
       else
-        format.html { render :action => "edit" }
-        format.xml  { render :xml => @mix.errors, :status => :unprocessable_entity }
+        format.html { render action: 'edit' }
+        format.json { render json: @mix.errors, status: :unprocessable_entity }
       end
     end
   end
 
   # DELETE /mixes/1
-  # DELETE /mixes/1.xml
+  # DELETE /mixes/1.json
   def destroy
-    @mix = Mix.find(params[:id])
     @mix.destroy
-
     respond_to do |format|
-      format.html { redirect_to(mixes_url) }
-      format.xml  { head :ok }
+      format.html { redirect_to mixes_url }
+      format.json { head :no_content }
     end
+  end
+
+  private
+  # Use callbacks to share common setup or constraints between actions.
+  def set_mix
+    @mix = Mix.find(params[:id])
+  end
+
+  # Never trust parameters from the scary internet, only allow the white list through.
+  def mix_params
+    params.require(:mix).permit(:name, :desc)
   end
 end
