@@ -1,8 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-Awesome::Application.config.session_store :cookie_store, :key => '_awesome_session'
-
-# Use the database for sessions instead of the cookie-based default,
-# which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
-# Awesome::Application.config.session_store :active_record_store
+Awesome4::Application.config.session_store :cookie_store, key: '_awesome4_session'
