@@ -1,9 +1,10 @@
-Awesome::Application.routes.draw do
-  resources :libraries
+Awesome4::Application.routes.draw do
+
+  root :to => "home#index"
 
   get "home/index"
 
-  resources :importages
+  resources :samples
 
   resources :mixes do
     member do
@@ -12,9 +13,11 @@ Awesome::Application.routes.draw do
       get 'add_sample/:sample_id' => 'mixes#add_sample'
     end
   end
-  
-  resources :samples
-  
+
+  resources :libraries
+
+  resources :importages
+
   devise_for :musicians
 
   resources :musicians do
@@ -23,23 +26,22 @@ Awesome::Application.routes.draw do
     end
   end
 
-  root :to => "home#index"
+  # The priority is based upon order of creation: first created -> highest priority.
+  # See how all your routes lay out with "rake routes".
 
-  # The priority is based upon order of creation:
-  # first created -> highest priority.
+  # You can have the root of your site routed with "root"
+  # root 'welcome#index'
 
-  # Sample of regular route:
-  #   match 'products/:id' => 'catalog#view'
-  # Keep in mind you can assign values other than :controller and :action
+  # Example of regular route:
+  #   get 'products/:id' => 'catalog#view'
 
-  # Sample of named route:
-  #   match 'products/:id/purchase' => 'catalog#purchase', :as => :purchase
-  # This route can be invoked with purchase_url(:id => product.id)
+  # Example of named route that can be invoked with purchase_url(id: product.id)
+  #   get 'products/:id/purchase' => 'catalog#purchase', as: :purchase
 
-  # Sample resource route (maps HTTP verbs to controller actions automatically):
+  # Example resource route (maps HTTP verbs to controller actions automatically):
   #   resources :products
 
-  # Sample resource route with options:
+  # Example resource route with options:
   #   resources :products do
   #     member do
   #       get 'short'
@@ -51,34 +53,24 @@ Awesome::Application.routes.draw do
   #     end
   #   end
 
-  # Sample resource route with sub-resources:
+  # Example resource route with sub-resources:
   #   resources :products do
   #     resources :comments, :sales
   #     resource :seller
   #   end
 
-  # Sample resource route with more complex sub-resources
+  # Example resource route with more complex sub-resources:
   #   resources :products do
   #     resources :comments
   #     resources :sales do
-  #       get 'recent', :on => :collection
+  #       get 'recent', on: :collection
   #     end
   #   end
 
-  # Sample resource route within a namespace:
+  # Example resource route within a namespace:
   #   namespace :admin do
   #     # Directs /admin/products/* to Admin::ProductsController
   #     # (app/controllers/admin/products_controller.rb)
   #     resources :products
   #   end
-
-  # You can have the root of your site routed with "root"
-  # just remember to delete public/index.html.
-  # root :to => "welcome#index"
-
-  # See how all your routes lay out with "rake routes"
-
-  # This is a legacy wild controller route that's not recommended for RESTful applications.
-  # Note: This route will make all actions in every controller accessible via GET requests.
-  # match ':controller(/:action(/:id(.:format)))'
 end
