@@ -1,7 +1,5 @@
 class ApplicationController < ActionController::Base
-  protect_from_forgery
-
-  def raiser(fo = nil)
-    raise fo.inspect
-  end
+  # Prevent CSRF attacks by raising an exception.
+  # For APIs, you may want to use :null_session instead.
+  protect_from_forgery with: :exception
 end
