@@ -1,5 +1,5 @@
 class Mix < ActiveRecord::Base
-belongs_to :musician
+  belongs_to :musician
   has_many :importages
   has_many :samples, :through => :importages
 end
