@@ -1,5 +1,5 @@
-# Load the rails application
+# Load the rails application.
 require File.expand_path('../application', __FILE__)
 
-# Initialize the rails application
-Awesome::Application.initialize!
+# Initialize the rails application.
+Awesome4::Application.initialize!
