@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Awesome::Application.config.secret_token = '5e87325b1c288e0f1b1c2dd69575cf73416783704a777366d673a4e9252e30c83f7d0395751d5f209d4ee5c84545bf3bccc75930098728434125e41bf2a9b693'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+Awesome4::Application.config.secret_key_base = '6c42e2ebde8ccad4d419c9f1259bb94fdb314ed66b96803aa44dc38eb4936344a2f3fde520fcf83003e021fc9ffe4f0e814d0fc72de80f1335adc80e9e8c8471'
